mod25/Vendas/src/services/ClienteService.java
package services;

import dao.generics.GenericService;
import domain.Cliente;

public class ClienteService extends GenericService<Cliente> implements IClienteService {
    @Override
    public Class<Cliente> getTipoClasse() {
        return Cliente.class;
    }

    @Override
    public void atualiarDados(Cliente entity, Cliente entityCadastrado) {

    }

    public ClienteService() {
        super();
    }

    @Override
    public Object excluir(Long id) {
        // Implementação da exclusão, se necessário.
        return null;
    }





    /*private IClienteDAO clienteDAO;
    public ClienteService(IClienteDAO clienteDAO) {
        this.clienteDAO = clienteDAO;
    }

    @Override
    public Boolean salvar(Cliente cliente) throws TipoChaveNaoEncontradaException {
        return clienteDAO.cadastrar(cliente);
    }

    @Override
    public Cliente buscarPorCPF(Long cpf) {
        return clienteDAO.consultar(cpf);
    }

    @Override
    public void alterar(Cliente cliente) throws TipoChaveNaoEncontradaException {
        clienteDAO.alterar(cliente);
    }*/
}
